сторінка друзі

+++ Social_network/user_app/services/friend_actions.py
@@ -38,8 +38,7 @@ def accept_friend_request(current_user, other_user):
         'friend': other_user
     }
 
-
-def delete_friendship(current_user, to_user):
+def any_delete(current_user, to_user):
     friendship = Friendship.objects.filter(
         from_user = current_user,
         to_user = to_user
@@ -52,9 +51,39 @@ def delete_friendship(current_user, to_user):
         ).first()
 
     if friendship:
-        friendship.delete()
-
+        if friendship.status == 'accepted' or friendship.status == "pending":
+            friendship.delete()
+        # else:
+        #     friendship.status = 'status'
+        #     friendship.save()
+    else:
+        Friendship.objects.get_or_create(
+            from_user = current_user,
+            to_user = to_user,
+            defaults = {'status': 'dismissed'}
+        )
+        
     return {
         'remove': True
     }
+
+
+# def delete_friendship(current_user, to_user):
+#     friendship = Friendship.objects.filter(
+#         from_user = current_user,
+#         to_user = to_user
+#     ).first()
+
+#     if not friendship:
+#         friendship = Friendship.objects.filter(
+#             from_user = to_user,
+#             to_user = current_user
+#         ).first()
+
+#     if friendship:
+#         friendship.delete()
+
+#     return {
+#         'remove': True
+#     }
